refactor(air): replace bitwise_pow2 with stand alone bitwise aux_table
    - remove pow2 air aux_tale
    - combine bitwise with bitwise_pow2 aux_table
    - replace bitwise_pow2 with new bitwise table in aux_table module
